+++ addons/frontend/index.html
@@ -107,7 +107,7 @@
       );
     </script>
     <script type="module" src="/src/main.js"></script>
-    <script type="module" src="/src/svg.ts"></script>
+    <script type="module" src="/src/typst.ts"></script>
   </head>
 
   <body
