+++ addons/frontend/src/typst-animation.ts
